perf: improve theme preview modal
